Add GNU General Public License version 3
This commit adds the GNU General Public License version 3 to the project. This license allows others to copy, distribute and modify the software. It makes clear that this is free software: users have the freedom to use, study, share, and modify the software. Incorporating the GNU GPLv3 in our project ensures that all contributors are aware of their rights and responsibilities when contributing to this project.